edit tests
Edit methods add and remove for correct work tests.

+++ orchestrator-model/src/main/java/model/entity/Container.java
@@ -29,7 +29,7 @@ public class Container extends ContainerBase {
 
 	@Override
 	public void setService(final Service service) {
-		if (hasService()) {
+		if (hasService() && getService().containsContainer(this)) {
 			getService().removeContainer(this);
 		}
